refec(GameManager): GameManager 패키지 이동
GameManager 패키지 이동:
- GameApplication 과의 위계 정리
- 게임 룰 역할을 하는 패키지는 rule 패키지에 넣음
